refactor: changes ids from usize to u64

+++ cli/src/file/delete.rs
@@ -3,7 +3,7 @@ use pcloud::http::HttpClient;
 
 #[derive(Parser)]
 pub struct Command {
-    file_id: usize,
+    file_id: u64,
 }
 
 impl Command {
